Follow button on ProjectCard working with refactored code
